\name{example}
\alias{example}
\docType{data}
\title{
Dataset used in example. 
}
\description{
A data frame \code{data} with 6000 observations on outcome, covariates, and 152 SNPs. A data frame \code{pathway} define the pathway. 
}
\usage{data(example, package = "ARTP3")}

\examples{
data(example)
head(data[, 1:6])
head(pathway)
}
\keyword{datasets}
