\name{geno}
\alias{geno}
\docType{data}
\title{
  A vector of file names used in example of \code{pathway.rawData}. 
}
\description{
  Names of 15 genotype files. Those files are plain text files, each stores genotypes data of a gene. 
}
\usage{data(geno)}

\examples{
  data(geno)
  head(geno)
}
\keyword{datasets}
