\name{ARTP3-package}
\alias{ARTP3-package}
\alias{ARTP3}
\docType{package}
\title{
Calculate gene and pathway p-values using the ARTP Test and summary data
}
\description{
Calculate gene and pathway p-values using the ARTP Test and summary data
}
\details{
\tabular{ll}{
Package: \tab ARTP3\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-05-26\cr
License: \tab GPL-2 + file LICENSE\cr
}
\code{\link{pathway.summaryData}}
}
\author{
Han Zhang, Bill Wheeler, Kai Yu, Yifan Yang

Maintainer: Han Zhang <han.zhang2@nih.gov>
}
\references{
Yu K, Li Q, Bergen AW, Pfeiffer RM, Rosenberg PS, Caporaso N, Kraft P, Chatterjee N. (2009) Pathway analysis by adaptive combination of P-values. 
Genet Epidemiol 33(8): 700 - 709
}

\keyword{pathway}

