\name{pathway.summaryData}
\alias{pathway.summaryData}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ARTP test for summary data}
\description{
Calculate gene and pathway p-values using the ARTP test and summary data
}
\usage{
pathway.summaryData(summary.files, pathway, reference, options = NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{summary.files}{
a character vector of file names containing the summary results of SNPs included in one or multiple studies. Each file must be able to be read by \code{\link[utils]{read.table}}. Each file must have columns called "SNP", "RefAllele", "EffectAllele", "BETA", and at least one of "SE", "P". An optional column called "N" for the non-missing sample size can also be included. See \code{Details}.
}
  \item{pathway}{
a character of the name of file containing definition of a pathway. It must be able to be read by \code{\link[utils]{read.table}} and have columns called "SNP", "Gene", "Chr".
}
  \item{reference}{
a data.frame containing the paths of binary PLINK files of reference dataset. It must have columns called "bed", "bim" and "fam". The current version allows user to specify multiple sets of PLINK files as separate rows, each contains genotypes and other informations of different SNPs (e.g., by chromosome).
}
  \item{options}{
a list of options to control the test procedure. If \code{NULL}, default options will be used. See \code{\link{options}}.
}
}
\details{
This function computes gene and pathway p-values when only summary data is available. Only the ARTP test modified from Yu et al. (2009) is well tested and is released with this package. ARTP is the Adaptive Rank Truncated Product test.

Each file in \code{summary.files} must contain
\itemize{
\item{\code{SNP}}{  SNP name}
\item{\code{RefAllele}}{  reference allele. Can be different in studies}
\item{\code{EffectAllele}}{  effect allele. Can be different in studies}
\item{\code{BETA}}{  estimated effect in linear regression model or log odds ratio in logistic regression model}
}
and must contain one of the optional columns
\itemize{
\item{\code{SE}}{  estimated standard error of BETA}
\item{\code{P}}{  p-value of Wald's, LRT or score test for testing H_0: BETA = 0. Can be generated by \code{\link[stats]{lm}}, \code{\link[stats]{glm}}, \code{\link[stats]{anova}} in \code{R} or other standard statistical softwares.}
}
Another optional column "N" is encouraged to be provided by the user
\itemize{
\item{\code{N}}{  non-missing sample size of SNP in a study}
}


The order of columns in files \code{summary.files}, \code{pathway} or in data frame \code{reference} are arbitrary, and all unnecessary columns (if any) are discarded in the analysis. 
}
\value{
\code{pathway.summaryData} returns an object of class \code{"ARTP3"}. It is a list containing the following components: 
\item{pathway.pvalue}{final pathway p-value accounting for multiple comparisons.}
\item{gene.pvalue}{a data frame containing gene name, number of SNPs in the gene that were included in the analysis, chromosome name, and the p-value for the gene accounting for multiple comparisons.}
\item{pathway}{a data frame defining pathway that was actually tested after various filters applied. }
\item{model}{a list containing detailed information of selected SNPs in each gene. }
\item{most.sig.genes}{a character vector of genes selected by \code{ARTP3}. They are the most promising candidates, although its statistical significance is not guaranteed. }
\item{deleted.snps}{a data frame containing SNPs excluded from the analysis and their reasons.}
\item{deleted.genes}{a data frame containing genes excluded from the analysis and their reasons. Set \code{options$rm.gene.subset} to be \code{FALSE} to include all genes. }
\item{options}{a list of options used in the analysis. See \code{\link{options}}}
\item{accurate}{\code{TRUE} if \code{options$nperm} is large enougth to accurately estimate p-values, i.e., if the criteria 
\code{sqrt(pvalue*(1-pvalue)/nperm)/pvalue < 0.1}
is satisfied.
}
\item{setup}{a list containing necessary input for \code{\link{pathway.warm.start}}. It can be written to a file by using the function \code{\link[base]{save}}, then its path can be the input of \code{\link{pathway.warm.start}}. Loading from \code{reference}, it also contains a data frame of genotypes of used SNPs (\code{setup$ref.geno}), if \code{options$keep.geno} is \code{TRUE}. }
}
\references{
Yu K, Li Q, Bergen AW, Pfeiffer RM, Rosenberg PS, Caporaso N, Kraft P, Chatterjee N. (2009) Pathway analysis by adaptive combination of P-values. 
Genet Epidemiol 33(8): 700 - 709

Zhang H, Shi J, Liang F, Wheeler W, Stolzenberg-Solomon R, Yu K.  (2014) A fast multilocus test with adaptive SNP selection for large-scale genetic association studies. European Journal of Human Genetics, 22, 696 - 702
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{options}}, \code{\link{pathway.warm.start}}, \code{\link{pathway.rawData}}
}
\examples{

library(ARTP3)

# Path of files containing summary statistics
# Only required columns will be loaded
study1 <- system.file("extdata", package = "ARTP3", "study1.txt.gz")
study2 <- system.file("extdata", package = "ARTP3", "study2.txt.gz")

# Path of a build-in file containing pathway definition
pathway <- system.file("extdata", package = "ARTP3", "pathway.txt.gz")

# Create data frame containing paths of build-in PLINK files that are going to used as reference
# As an example, use all chromosomes
chr <- 1:22
nchr <- length(chr)

fam <- vector("character", nchr)
bim <- vector("character", nchr)
bed <- vector("character", nchr)

for(i in 1:nchr){
  fam[i] <- system.file("extdata", package = "ARTP3", paste("1000genomes.chr", chr[i], ".fam", sep = ""))
  bim[i] <- system.file("extdata", package = "ARTP3", paste("1000genomes.chr", chr[i], ".bim", sep = ""))
  bed[i] <- system.file("extdata", package = "ARTP3", paste("1000genomes.chr", chr[i], ".bed", sep = ""))
}

reference <- data.frame(fam, bim, bed)

# Set the options. 
# Accumulate signal from the top 5 SNPs in each gene
# 1e5 replicates of resampling to estimate the p-value
options <- list(inspect.snp.n = 5, nperm = 1e5, 
                maf = .01, HWE.p = 1e-6, 
                gene.R2 = .9, 
                id.str = "unique-pathway-id", 
                out.dir = getwd(), save.setup = FALSE)

## pathway test with two study files
# ret <- pathway.summaryData(summary.files = c(study1, study2), pathway, reference, options = options)
# 
# ret$pathway.pvalue
## [1] 0.00696493  # mac os
## [1] 0.007394926 # Windows
# head(ret$gene.pvalue)

options$print <- FALSE # turn off printing

## pathway test with each of study files
# ret1 <- pathway.summaryData(summary.files = study1, pathway, reference, options = options)
# ret2 <- pathway.summaryData(summary.files = study2, pathway, reference, options = options)
# 
# ret1$pathway.pvalue
## [1] 0.08872411 # Mac OS
## [1] 0.0899991  # Windows
# ret2$pathway.pvalue
## [1] 0.05579944 # Mac OS
## [1] 0.05476945 # Windows


##################################################
## The reference genotype data can also be merged into a single set of PLINK files
bed <- system.file("extdata", package = "ARTP3", "1000genomes.bed")
bim <- system.file("extdata", package = "ARTP3", "1000genomes.bim")
fam <- system.file("extdata", package = "ARTP3", "1000genomes.fam")

reference <- data.frame(fam, bim, bed)
# ret.comb <- pathway.summaryData(summary.files = c(study1, study2), pathway, reference, options = options)
## result should be the same as before
# ret.comb$pathway.pvalue == ret$pathway.pvalue

}
